Adjust the consume API to explicitly reject non-iterable-of-dicts
As it stood the API would accept anything and pass it on where it
eventually would fail to quack like a duck and crash in some interesting
way. This just checks it's an iterable of dictionaries and raises a
ValueError if it's not so programmers don't have to dig into the duck
pond.

Signed-off-by: Jeremy Cline <jcline@redhat.com>